Move methods into character.cpp/character_status.cpp/adventurer.cpp
