+++ casino.cpp
@@ -4,6 +4,7 @@
 #include "calc.hpp"
 #include "casino_card.hpp"
 #include "character.hpp"
+#include "character_status.hpp"
 #include "config.hpp"
 #include "elona.hpp"
 #include "i18n.hpp"
