+++ character_status.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace elona
+{
+
+void modcorrupt(int = 0);
+void modify_potential(int, int, int);
+void modify_karma(int, int);
+void modheight(int = 0, int = 0);
+void modweight(int, int, bool = false);
+void refreshspeed(int = 0);
+void refresh_speed_correction_value();
+void gain_new_body_part(int);
+void gain_level(int);
+void label_1455(int);
+void label_1456(int);
+
+}
