+++ equipment.hpp
@@ -1,0 +1,17 @@
+#pragma once
+#include <string>
+
+namespace elona
+{
+
+void equipinfo(int = 0, int = 0, int = 0);
+int eqweaponheavy();
+int eqweaponlight();
+void colorres(int = 0);
+void eqrandweaponmage(int = 0);
+void wear_most_valuable_equipment_for_all_body_parts();
+void wear_most_valuable_equipment();
+void supply_new_equipment();
+void label_1530();
+
+} // namespace elona
