+++ init.cpp
@@ -1,4 +1,5 @@
 #include "ability.hpp"
+#include "adventurer.hpp"
 #include "autopick.hpp"
 #include "blending.hpp"
 #include "buff.hpp"
@@ -824,7 +825,7 @@ void initialize_game()
         initlv = 50;
         rc = 0;
         flt(100);
-        characreate(0, 84, -3, 0);
+        chara_create(0, 84, -3, 0);
         initialize_pc_character();
         gdata_year = 517;
         gdata_month = 12;
@@ -849,7 +850,7 @@ void initialize_game()
         cdata[0].platinum_coin = 30;
         cdata[0].fame = 65000;
         gdata_main_quest_flag = 100;
-        refresh_character(0);
+        chara_refresh(0);
         flt();
         itemcreate(0, 284, -1, -1, 0);
         inv[ci].number = 20;
