+++ itemgen.cpp
@@ -3,6 +3,7 @@
 #include "calc.hpp"
 #include "character.hpp"
 #include "enchantment.hpp"
+#include "i18n.hpp"
 #include "item.hpp"
 #include "item_db.hpp"
 #include "item_material.hpp"
@@ -681,7 +682,7 @@ void change_item_material()
     }
     apply_item_material();
     label_1583();
-    refresh_character(cc);
+    chara_refresh(cc);
     return;
 }
 
@@ -775,4 +776,22 @@ void set_material_specific_attributes()
     }
     return;
 }
+
+void add_quality_parentheses()
+{
+    if (fixlv == 4)
+    {
+        cdatan(0, rc) = i18n::_(u8"ui", u8"bracket_left") + cdatan(0, rc)
+            + i18n::_(u8"ui", u8"bracket_right");
+        cdata[rc].level = cdata[rc].level * 10 / 8;
+    }
+    if (fixlv == 5)
+    {
+        cdatan(0, rc) =
+            lang(u8"《"s, u8"{"s) + cdatan(0, rc) + lang(u8"》"s, u8"}"s);
+        cdata[rc].level = cdata[rc].level * 10 / 6;
+    }
+    return;
+}
+
 } // namespace elona
