+++ variables.hpp
@@ -666,8 +666,6 @@ ELONA_EXTERN(std::string userfile);
 ELONA_EXTERN(std::string usermsg);
 bool actionsp(int, int);
 int addbuff(int = 0, int = 0, int = 0, int = 0);
-int advfavoriteskill(int = 0);
-int advfavoritestat(int = 0);
 int breath_list();
 int calcincome(int = 0);
 int calcmagiccontrol(int = 0, int = 0);
@@ -677,16 +675,14 @@ int cardcanblock(int = 0);
 int cardcandeclareattack(int = 0);
 int cardcanuseskill(int = 0);
 int cargocheck();
-int carmor(int = 0);
 int cbreeder(int = 0);
 int cdbit(int = 0, int = 0);
-bool chara_unequip(int);
-int characreate(int = 0, int = 0, int = 0, int = 0);
 int clientguide(int = 0);
+void initialize_character();
+void initialize_pc_character();
+void chara_preparepic(int = 0, int = 0);
 int convertartifact(int = 0, int = 0);
-int copy_chara(int = 0);
 int create_card(int = 0, int = 0);
-int customtalk(int = 0, int = 0);
 int direction(int = 0, int = 0, int = 0, int = 0);
 int discsetmc();
 int dist(int = 0, int = 0, int = 0, int = 0);
@@ -697,17 +693,11 @@ void damage_insanity(int, int);
 void dmgsp(int, int);
 int efstatusfix(int = 0, int = 0, int = 0, int = 0);
 int eleinfo(int = 0, int = 0);
-int eqweaponheavy();
-int eqweaponlight();
 int exist_questtarget();
-int findally(int = 0);
 int findbuff(int = 0, int = 0);
-int findchara(int = 0);
 int findunid(const std::string&);
 int fltsetdungeon();
 int gameover();
-int get_freeally();
-int get_freechara();
 int getdecksum(int = 0);
 int getholdersum(int = 0);
 void getinheritance(int, elona_vector1<int>&, int&);
@@ -716,7 +706,6 @@ int getunid(int = 0);
 int getworker(int = 0, int = 0);
 int iequiploc(int = 0);
 int imeget();
-int implevel(int = 0);
 int key_direction();
 int access_item_db(int);
 int voting_box();
@@ -725,7 +714,6 @@ int try_to_reveal();
 int can_evade_trap();
 int try_to_disarm_trap();
 int try_to_perceive_npc(int);
-int do_create_character();
 int initialize_world_map();
 int label_1753();
 int get_card_info();
@@ -767,7 +755,6 @@ int label_2230();
 int label_2231();
 bool talk_give_potion_of_cure_corruption();
 int new_ally_joins();
-int label_2664();
 int label_2672();
 int label_2694();
 int label_2695();
@@ -793,7 +780,6 @@ int random_material(int = 0, int = 0);
 int randomele();
 int randskill();
 int relationbetween(int, int);
-int relocate_chara(int = 0, int = 0, int = 0);
 int roll(int, int, int);
 int roll_max(int, int, int);
 int roundmargin(int = 0, int = 0);
@@ -863,7 +849,6 @@ std::string npcn(int = 0);
 std::string random_title(int = 0);
 std::string randomname();
 std::string ranktitle(int = 0);
-std::string refchara_str(int = 0, int = 0);
 std::string sncnv(const std::string&);
 std::string trim_item_description(const std::string&, bool);
 std::string txtbuilding(int = 0, int = 0);
@@ -874,9 +859,6 @@ std::string yourself(int = 0);
 void actionproc();
 void addbuilding(int = 0, int = 0, int = 0, int = 0);
 void addefmap(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
-void addnews(int = 0, int = 0, int = 0, const std::string& = "");
-void addnews2(const std::string&, int = 0);
-void addnewstopic(const std::string&, const std::string&);
 void anime_halt();
 void animeblood(int = 0, int = 0, int = 0);
 void animeload(int = 0, int = 0);
@@ -893,13 +875,10 @@ void cardhelp(const std::string&, int = 0);
 void cardpos(int = 0, int = 0);
 void cdbitmod(int = 0, int = 0, int = 0);
 void cell_draw();
-void chara_preparepic(int = 0, int = 0);
-void chara_vanquish(int = 0);
 void check_kill(int = 0, int = 0);
 void check_quest();
 void cnv_filestr(std::string&);
 void cnvbonus(int = 0, int = 0);
-void colorres(int = 0);
 void cpflip();
 void cpisenemy();
 void cpisme();
@@ -909,15 +888,12 @@ void cslinedown();
 void cslineup();
 void cursor_check();
 void cutname(std::string&, int = 0);
-void del_chara(int = 0);
 void delbottomcard(int = 0);
 void delbuff(int = 0, int = 0);
 void dipcursed(int = 0, int = 0);
 void dmgcard(int = 0, int = 0);
 void dmgplayer(int = 0, int = 0);
 void efllistadd(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
-void eqrandweaponmage(int = 0);
-void equipinfo(int = 0, int = 0, int = 0);
 void exittempinv();
 void fileadd(const fs::path& filepath, int = 0);
 void fix_input_chat(std::string&);
@@ -973,11 +949,6 @@ void label_1346();
 void end_dmghp();
 void initialize_server_info();
 void label_1399();
-void refresh_speed_correction_value();
-void gain_new_body_part(int);
-void gain_level(int);
-void label_1455(int);
-void label_1456(int);
 void label_1457();
 void label_1458();
 void label_1459(int);
@@ -993,18 +964,13 @@ void label_1469(int);
 void label_1470(int);
 void label_1471(int);
 void label_1472(int);
-void refresh_character(int);
 void label_1512(int);
 void gain_disarm_trap_experience();
 void get_hungry(int);
 void label_1520(int);
 void refresh_burden_state();
 void gain_special_action();
-void label_1530();
-void place_character();
 void add_quality_parentheses();
-void initialize_character();
-void initialize_pc_character();
 void label_1537();
 void label_1538();
 void label_15380();
@@ -1202,13 +1168,6 @@ void label_2266();
 void label_2267();
 void initialize_economy();
 void get_random_npc_id();
-void create_all_adventurers();
-void create_adventurer();
-void label_2662();
-void set_item_which_will_be_used();
-void wear_most_valuable_equipment_for_all_body_parts();
-void wear_most_valuable_equipment();
-void supply_new_equipment();
 void label_2669();
 void label_2670();
 void label_2671();
@@ -1226,6 +1185,8 @@ void initialize_game();
 bool turn_wrapper();
 turn_result_t pass_turns(bool);
 turn_result_t turn_begin();
+void lovemiracle(int = 0);
+void resistmod(int = 0, int = 0, int = 0);
 void label_2735();
 void label_2736();
 void label_27412();
@@ -1236,18 +1197,11 @@ void play_the_last_scene_again();
 turn_result_t pc_died();
 void show_game_score_ranking();
 void lenfix(std::string&, int = 0);
-void lovemiracle(int = 0);
 void make_sound(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void makecardlist();
 void matdelmain(int = 0, int = 0);
 void matgetmain(int = 0, int = 0, int = 0);
-void modcorrupt(int = 0);
-void modify_potential(int, int, int);
-void modheight(int = 0, int = 0);
-void modimp(int = 0, int = 0);
-void modify_karma(int, int);
 void modrank(int = 0, int = 0, int = 0);
-void modweight(int, int, bool = false);
 void msg_clear();
 void msg_halt();
 void msg_newline();
@@ -1257,11 +1211,9 @@ void page_load();
 void page_save();
 void press(bool = false);
 void put_questtarget();
-void refreshspeed(int = 0);
 void remain_make(int = 0, int = 0);
 void removeitem(int, int);
 void removeworker(int = 0);
-void resistmod(int = 0, int = 0, int = 0);
 void ride_begin(int = 0);
 void ride_end();
 void rm_crlf(std::string&);
@@ -1313,7 +1265,6 @@ void windowanime(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void windowanimecorner(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void cs_listbk();
 void cs_list(bool, const std::string&, int, int, int = 0, int = 0, int = 0);
-int access_character_info();
 
 template <typename T>
 void txt()
